Change Motion tween to have a "object" property.
The x and y values computed by the Tween will be automatically applied to the
object's x and y properties too. The object can have any type, as long as it
allows assignment to its x and y properties (so instances of Entity, Point,
or Image will all work). The change is backwards compatible, and can be used
as before regardless of whether the object is set.